Do not alter the order of genes.
If the order of the genes on targets differ from the order in data
some genes are lost and the analysis is wrong.

The correct approach to sort genes is using `sort -V`
which sorts the alphabetic part and sorts naturally the numbers.

However, using that approach needs a preprocessing
which will not be implemented today.

So, for now just do a correct analysis.